# Conflicts:
#	feature/navigator/src/main/java/com/record/navigator/MainNavigator.kt
#	feature/video/src/main/java/com/record/video/navigation/VideoNavigation.kt

+++ data/video/src/main/java/com/record/video/model/remote/response/RecordInfo.kt
@@ -12,11 +12,11 @@ data class RecordInfo(
     @SerialName("fileUrl")
     val fileUrl: FileUrl,
     @SerialName("id")
-    val id: Int,
+    val id: Long,
     @SerialName("location")
     val location: String,
     @SerialName("uploaderId")
-    val uploaderId: Int,
+    val uploaderId: Long,
     @SerialName("uploaderNickname")
     val uploaderNickname: String,
     @SerialName("isMine")
